update annotation & package name
