+++ research/gaia-x/ir/compiler/src/test/java/com/alibaba/graphscope/common/TestUtils.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.alibaba.grahscope.common;
+package com.alibaba.graphscope.common;
 
 import org.apache.commons.io.FileUtils;
